pyproject.toml
[project]
name = "the-datastar-way"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "brotli-asgi>=1.5.0",
    "datastar-py>=0.7.0",
    "python-fasthtml>=0.12.35",
]

[tool.ruff.lint]
ignore = ["F401", "F403", "F405"]
